# HIP: improve perf for quantized embedding forward kernel