--- a/fbgemm_gpu/codegen/embedding_forward_quantized_split_template.cu
+++ b/fbgemm_gpu/codegen/embedding_forward_quantized_split_template.cu
@@ -294,7 +294,11 @@ __global__ void {{ type_map[emb_weight_type].enum_name }}_split_embedding{{ "_no
     // equivalent to fence + wait.
     cp_async_wait<0>();
 #ifdef __HIP_PLATFORM_HCC__
-    __syncthreads();
+    // performance: replace global __syncthreads with per CU __threadfence_block
+    // __threadfence_block is perfectly fine replacement for __syncwarp, because
+    // a. memory fencing: __threadfence_block ops. at CU level, same as __syncwarp at SM
+    // b. thread re-converge: __syncwarp at scope of warp; AMD wave executes in lockstep
+    __threadfence_block();
 #else
     __syncwarp();
 #endif